refactor: product item refactor
